requirements.txt
torch
torchvision
numpy
pillow
uvicorn
fastapi
scikit-image